Fix copier settings to exclude template metadata (#4)
Files in the top-level directory of the template were getting copied
because there was not a setting that indidcated only the template
sub-folder should be copied.

This fixes that by adding a _subdirectory setting to copier.yml.

Additionally, a `template` subdirectory has been added to allow
for templated package directory names.